added test for image resolution and object storage endpoint
